requirements.txt
dash
dash-cytoscape
requests
pyyaml
networkx